requirements.txt
# Cuttlefish Multi-Agent RAG System - Python Dependencies
# Generated from notebook analysis and reorganization

# ===========================================
# AI/ML and Language Models
# ===========================================

# LangChain ecosystem
langchain-core==0.2.38
langchain-openai==0.1.22
langchain-community==0.2.16
langchain-cohere==0.1.9
langchain-experimental==0.0.64
langgraph==0.2.14

# OpenAI
openai>=1.40.6

# ===========================================
# Web Framework and API
# ===========================================

# FastAPI (replacing Flask)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Original Flask dependencies (for reference/migration)
flask==3.0.0
flask-cors==4.0.0

# Pydantic for data validation (compatible with langchain-core)
pydantic>=2.7.4,<3.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
sqlalchemy>=1.4.0,<2.1.0

# ===========================================
# Database and Storage
# ===========================================

# Supabase integration
supabase>=2.2.0
postgrest>=0.13.0

# ===========================================
# Data Processing and Utilities
# ===========================================

# Data processing
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# ===========================================
# Additional Production Dependencies
# ===========================================

# Logging and monitoring
structlog==23.2.0

# HTTP requests
httpx>=0.24.0,<1.0.0
requests>=2.31.0

# Web Search API
tavily-python>=0.3.0

# Google Cloud Platform (for LogSearch agent)
google-cloud-logging>=3.8.0

# Async support (built into Python 3.7+)
# asyncio is part of standard library

# ===========================================
# Development and Testing (Optional)
# ===========================================

# Uncomment for development
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0
# jupyter>=1.0.0

# ===========================================
# Optional Performance Enhancements
# ===========================================

# PostgreSQL database driver (required for Supabase PostgreSQL)
psycopg2-binary>=2.9.0  # For PostgreSQL connections

# Uncomment for better performance
# asyncpg>=0.28.0  # Async PostgreSQL driver  
# orjson>=3.9.0  # Fast JSON parsing

# ===========================================
# Advanced RAG Ensemble Dependencies
# ===========================================

# For advanced ensemble retrieval features
cohere>=4.50.0  # For ContextualCompressionRetriever reranking (optional)